Fixed duration in wait_for_server

+++ wasm-fuzzer/fuzzing-client-afl/socket_client.h
@@ -16,12 +16,14 @@
 #include <algorithm>
 #include <ctime>
 #include <cstdio>
+#include <chrono>
+#include <thread>
 
 void error(const char *);
 int connectToServer(char *, int);
 void clientWrite(int, char *, int);
 void clientRead(int, char *, int);
-void wait_for_server(char *, int, int = 4000, double = 40000);
+void wait_for_server(char *, int, int = 4000, int = 40000);
 void runClient(int, char *, int, char *, char *, int);
 
 #endif
